# openapi.yaml
openapi: 3.0.3
info:
  title: GlobeCo Trade Blotter API
  version: 1.0.0
  description: |
    API for managing trade orders, blotters, blocks, allocations, and trades for the GlobeCo Suite.
  contact:
    name: Noah Krieger
    email: noah@kasbench.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /blotter/{blotterId}:
    get:
      tags: [Blotter]
      summary: Get a blotter by ID
      parameters:
        - in: path
          name: blotterId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Blotter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blotter'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Blotter]
      summary: Update a blotter
      parameters:
        - in: path
          name: blotterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blotter'
      responses:
        '200':
          description: Updated blotter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blotter'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Blotter]
      summary: Delete a blotter
      parameters:
        - in: path
          name: blotterId
          required: true
          schema:
            type: integer
        - in: query
          name: versionId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /order/{orderId}/blotter/{blotterId}:
    post:
      tags: [Order]
      summary: Update the blotterId for an order
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
        - in: path
          name: blotterId
          required: true
          schema:
            type: integer
        - in: query
          name: versionId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    put:
      tags: [Order]
      summary: Update the blotterId for an order
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
        - in: path
          name: blotterId
          required: true
          schema:
            type: integer
        - in: query
          name: versionId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /block/{blockId}:
    get:
      tags: [Block]
      summary: Get a block by ID
      parameters:
        - in: path
          name: blockId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Block
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Block]
      summary: Update a block
      parameters:
        - in: path
          name: blockId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockRequestDTO'
      responses:
        '200':
          description: Updated block
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Block]
      summary: Delete a block
      parameters:
        - in: path
          name: blockId
          required: true
          schema:
            type: integer
        - in: query
          name: versionId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /trade:
    get:
      tags: [Trade]
      summary: Get all trades
      responses:
        '200':
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trade'
    post:
      tags: [Trade]
      summary: Create a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeRequestDTO'
      responses:
        '201':
          description: Created trade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
  /trade/{tradeId}:
    get:
      tags: [Trade]
      summary: Get a trade by ID
      parameters:
        - in: path
          name: tradeId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Trade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Trade]
      summary: Update a trade
      parameters:
        - in: path
          name: tradeId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeRequestDTO'
      responses:
        '200':
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Trade]
      summary: Delete a trade
      parameters:
        - in: path
          name: tradeId
          required: true
          schema:
            type: integer
        - in: query
          name: versionId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /trade/block/{blockId}:
    get:
      tags: [Trade]
      summary: Get all trades for a block
      parameters:
        - in: path
          name: blockId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trade'
  /trade/{tradeId}/fill/{quantityFilled}:
    post:
      tags: [Trade]
      summary: Fill a trade
      parameters:
        - in: path
          name: tradeId
          required: true
          schema:
            type: integer
        - in: path
          name: quantityFilled
          required: true
          schema:
            type: number
            format: double
        - in: query
          name: versionId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /trade/allocateProRata:
    post:
      tags: [Trade]
      summary: Pro rata allocation of trade.filledQuantity to block allocations of the referenced block
      parameters:
        - in: query
          name: tradeId
          required: true
          schema:
            type: integer
        - in: query
          name: versionId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated trade with pro rata allocation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  schemas:
    SecurityType:
      type: object
      properties:
        id:
          type: integer
        abbreviation:
          type: string
        description:
          type: string
        version:
          type: integer
      required: [id, abbreviation, description, version]
    Blotter:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        autoPopulate:
          type: boolean
        securityTypeId:
          type: integer
          nullable: true
        version:
          type: integer
      required: [id, name, autoPopulate, version]
    TradeType:
      type: object
      properties:
        id:
          type: integer
        abbreviation:
          type: string
        description:
          type: string
        version:
          type: integer
      required: [id, abbreviation, description, version]
    Destination:
      type: object
      properties:
        id:
          type: integer
        abbreviation:
          type: string
        description:
          type: string
        version:
          type: integer
      required: [id, abbreviation, description, version]
    OrderType:
      type: object
      properties:
        id:
          type: integer
        abbreviation:
          type: string
        description:
          type: string
        version:
          type: integer
      required: [id, abbreviation, description, version]
    OrderStatus:
      type: object
      properties:
        id:
          type: integer
        abbreviation:
          type: string
        description:
          type: string
        version:
          type: integer
      required: [id, abbreviation, description, version]
    Security:
      type: object
      properties:
        id:
          type: integer
        ticker:
          type: string
        description:
          type: string
        securityType:
          $ref: '#/components/schemas/SecurityType'
        version:
          type: integer
      required: [id, ticker, securityType, version]
    Order:
      type: object
      properties:
        id:
          type: integer
        security:
          $ref: '#/components/schemas/Security'
        blotter:
          $ref: '#/components/schemas/Blotter'
          nullable: true
        quantity:
          type: number
          format: double
        orderTimestamp:
          type: string
          format: date-time
        orderType:
          $ref: '#/components/schemas/OrderType'
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
          nullable: true
        version:
          type: integer
      required: [id, security, quantity, orderType, version]
    Block:
      type: object
      properties:
        id:
          type: integer
        security:
          $ref: '#/components/schemas/Security'
        orderType:
          $ref: '#/components/schemas/OrderType'
        version:
          type: integer
      required: [id, security, orderType, version]
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        block:
          $ref: '#/components/schemas/Block'
        quantity:
          type: number
          format: double
        tradeType:
          $ref: '#/components/schemas/TradeType'
        destination:
          $ref: '#/components/schemas/Destination'
          nullable: true
        filledQuantity:
          type: number
          format: double
        version:
          type: integer
      required: [id, block, quantity, filledQuantity, version]
...
    TradeRequestDTO:
      type: object
      properties:
        blockId:
          type: integer
        quantity:
          type: number
          format: double
        tradeTypeId:
          type: integer
        destinationId:
          type: integer
        filledQuantity:
          type: number
          format: double
        version:
          type: integer
      required: [blockId, quantity, version]
    BlockRequestDTO:
      type: object
      properties:
        securityId:
          type: integer
        orderTypeId:
          type: integer
        version:
          type: integer
      required: [securityId, orderTypeId]
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Conflict:
      description: Optimistic locking failure
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string 